code/requirements.txt
decorator==4.0.10
Django==1.10.1
django-model-utils==2.6
django-redis-sessions==0.5.6
functools32==3.2.3.post2
gitdb2==2.0.2
GitPython==2.1.5
jsonfield==1.0.3
jsonschema==2.5.1
networkx==1.11
pbr==1.10.0
publicsuffix==1.1.0
pyaml==15.8.2
pymongo==3.4.0
python-dateutil==2.6.0
PyYAML==3.12
redis==2.10.5
requests==2.12.4
six==1.10.0
smmap2==2.0.3
url==0.2.0
wheel==0.24.0
